Add initial newlines for Compact mode users
In compact mode, the first line is on the same line as the username,
distorting the grid.

This change will result in a blank line between the two messages for
Cozy mode users, but at least the experience will be more or less the
same for both since the second message gets split by the bot's name in
Compact mode as well.